% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-mutate.R
\name{loop_mutate}
\alias{loop_mutate}
\title{mutate in loop}
\usage{
loop_mutate(.data, idx, ...)
}
\arguments{
\item{.data}{a data.frame}

\item{idx}{idx vector}

\item{...}{expressions}
}
\description{
mutate in loop
}
\examples{
suppressMessages(library(dplyr))
# Looping trough exponents
cars \%>\%
  loop_mutate(2:4, paste0("speed.", .idx) := speed ** .idx) \%>\%
  head(5)
# Working on groups
}
