% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looplyr.R
\docType{package}
\name{looplyr-package}
\alias{looplyr-package}
\alias{looplyr}
\title{looplyr: Loop Verbs for A Grammar of Data Manipulation}
\description{
Provides a small set of tools for running
data manipulation in loop.
}
\author{
Krzysztof Joachimiak
}
\keyword{package}
