% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-mutate.R
\name{loop_mutate}
\alias{loop_mutate}
\title{Mutate in loop}
\usage{
loop_mutate(.data, .x, ...)
}
\arguments{
\item{.data}{a data.frame-like object}

\item{.x}{a vector the function is looping on}

\item{...}{Name-value pairs. The name gives the name of the column in the output.
The value can be:
\itemize{
\item A vector of length 1, which will be recycled to the correct length.
\item A vector the same length as the current group (or the whole data frame if ungrouped).
\item NULL, to remove the column.
\item A data frame or tibble, to create multiple columns in the output.
}}
}
\description{
Mutate in loop
}
\examples{
suppressMessages(library(dplyr))
# Looping trough exponents
cars \%>\%
  loop_mutate(2:4, paste0("speed.", .x) := speed ** .x) \%>\%
  head(5)
}
